pyproject.toml
[project]
name = "repl-nix-sillysopapillaaccess"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.3",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "oauthlib>=3.2.2",
    "requests>=2.32.3",
    "werkzeug>=3.0.6",
    "psycopg2-binary>=2.9.10",
    "flask-mail>=0.10.0",
    "pyjwt",
]
